graphic: Build one big list of vertices
This means less function calls to TurtleScreen.draw_line, but it also
means that each Point has now the color assigned and it's not passed
around as a uniform to the shader.

The scale-matrix is now also built in draw_and_update, which means more
DRY and thus a better program!