graphic: preserve order of drawings
In the original implementation, we had a clear order:
floodfills were in the background, lines in the middle and text on top
of everything.

This is probably not the way we want it, we want to preserve the order
in which things are drawn. This patch changes it: If you first draw a
text and then a line over it, the line will be in front of the text, as
expected.