graphic: add intermediate render struct
The frame buffer for the headless renderer has a fixed size, so when we
used it to draw on it with .draw(), we got an unexpected size. But for
windows, we still want to use the framebuffer for speed and because it
automatically resizes with the window.

This patch adds an intermediate struct, which for windows draws on the
window buffer and for headless renderers draws on a new texture that has
the right size. This way we get the correct image when we use
.sceenshot(), which is probably good for various reasons.