parser: get rid of the function table
We parse every function as some kind of "polish notation" now, like a reversed
stack language. This has the benefit that the parser does not need to keep
track of the functions and their argument counts, and it means that function
definitions are not block-scoped anymore, they can be function scoped just like
variables.

The drawback is that argument and function-not-found errors are catched at
runtime rather than compile time, but that is a small price to pay.

The implementation is not the cleanest, so I'm not entirely sure if the code
really got better that way, but it's a start.